feat(cq): more tuning up
Ref: #1040
