--- packages/api/src/external/carequality/command/cq-directory/shared.ts
@@ -1,3 +1,0 @@
-export const cqDirectoryEntryTemp = `cq_directory_entry_temp`;
-export const cqDirectoryEntry = `cq_directory_entry`;
-export const cqDirectoryEntryBackup = `cq_directory_entry_backup`;
